+++ Gateway/Response/RefundResponseHandler.php
@@ -1,0 +1,42 @@
+<?php
+/**
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is provided with Magento in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * Copyright © 2020 MultiSafepay, Inc. All rights reserved.
+ * See DISCLAIMER.md for disclaimer details.
+ *
+ */
+
+declare(strict_types=1);
+
+namespace MultiSafepay\ConnectCore\Gateway\Response;
+
+use Magento\Payment\Gateway\Helper\SubjectReader;
+use Magento\Payment\Gateway\Response\HandlerInterface;
+use Magento\Sales\Api\Data\OrderPaymentInterface;
+
+class RefundResponseHandler implements HandlerInterface
+{
+
+    /**
+     * @inheritDoc
+     */
+    public function handle(array $handlingSubject, array $response)
+    {
+        $paymentDataObject = SubjectReader::readPayment($handlingSubject);
+
+        /** @var OrderPaymentInterface $payment */
+        $payment = $paymentDataObject->getPayment();
+
+        $payment->setTransactionId($response['refund_id']);
+        $payment->setIsTransactionClosed(true);
+
+        return $this;
+    }
+}