+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added MultiSafepay refund_id to the refund transaction in the Magento Payment.
+
 ### Fixed
 - Added dependencies in module.xml and composer.json